core/requirements.txt
PyYAML
Jinja2
click
requests
semver
marshmallow
